+++ HWJavaBootCamp2024/src/main/java/com/colvir/accountant/dto/GenerateEmpRequest.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class GenerateEmpRequest {
 
-    private Long   idDepartment;
+    private Integer   idDepartment;
     private String surname;
     private String name;
     private String patronymic;
